# swagger_go/out/go-gin-server/go/api_store.go
/*
 * Swagger Petstore
 *
 * This is a sample server Petstore server.  You can find out more about     Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).      For this sample, you can use the api key `special-key` to test the authorization     filters.
 *
 * API version: 1.0.0
 * Contact: apiteam@swagger.io
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

import (
	"net/http"

	"github.com/gin-gonic/gin"
)

// DeleteOrder - Delete purchase order by ID
func DeleteOrder(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}

// GetInventory - Returns pet inventories by status
func GetInventory(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}

// GetOrderById - Find purchase order by ID
func GetOrderById(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}

// PlaceOrder - Place an order for a pet
func PlaceOrder(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}